Moved context integration tests to their own test class
